as of 3.2.13+ "vue/compiler-sfc" no longer explicitly required
Note: as of 3.2.13+, this package is included as a dependency of the main vue package and can be accessed as vue/compiler-sfc. This means you no longer need to explicitly install this package and ensure its version match that of vue's. Just use the main vue/compiler-sfc deep import instead.

https://www.npmjs.com/package/@vue/compiler-sfc